Move struct Message to from package process(io) to server(io)
These are server details, now they have higher cohesion.
